refactor(mcp): split the tool surface into per-group modules
server.ts held 2,391 LOC of inline tool registrations, which is why #2379
had to pin the newly-enforced backend file-size ceiling at 2,400 instead of
the frontend's 2,200. Move every registration into packages/backend/src/lib/
mcp/tools/, one module per tool group, and drop the temporary pin so both
roots share MAX_FILE_LOC again.

Each group module exports registerXTools({ server }) and receives the SAME
safety-wrapped registrar createMcpServer already built, so a tool cannot be
registered around the one-shot binding / scope gate / mutation-exec guards /
approval gate / snapshot-audit-notify flow. server.ts keeps exactly that
cross-cutting layer plus the scope-filtered tools/list view; the policy
tables it reads move to toolPolicy.ts so a group module can resolve a tool's
tier without importing server.ts and closing a cycle. TOOL_SCOPES,
tokenHasScope, MCP_KERNEL_TOOLS and isToolVisibleForScopes are re-exported
from server.ts, so no importer changes.

Pure mechanical extraction: the serialized tools/list payload — all 62 tool
names, descriptions and JSON schemas — is byte-identical to the pre-split
server, verified by dumping it from both trees and diffing.

server.ts 2,391 -> 382 LOC; largest backend file is now network/service.ts
at 2,046. New server.toolRegistry.test.ts pins the registered set against
TOOL_SCOPES so a dropped registerXTools() call reds instead of silently
deleting a whole tool group. assist_consistency's tool-name-drift scan
followed the registrations to tools/ — it was statically parsing server.ts
and would otherwise have gone vacuous.

Closes #2384
